# lib/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: nilsbrugger.nb.nb.nb@gmail.com
    name: Nils
  description: The time planer API
  license:
    name: Apache 2.0
  title: Time Planer
  version: 2.2.1-alpha
servers:
- url: http://eds.logfro.de/time-planer
- url: http://localhost:8080/time-planer
...
paths:
...
  /user/{username}/tasks:
    delete:
      description: Deletes all Tasks
      operationId: removeAllTasks
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: all tasks deleted
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: all tasks deleted
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Remove All Tasks
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
    get:
      description: Fetch the users tasks
      operationId: getAllTasks
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                  - name: do
                    deadline: 1944-12-07T00:00:00.000Z
                    planedDate: 2010-08-11T00:00:00.000Z
                    description: minim sunt in do culpa
                    importance: 0
                    entererAt: 1988-10-06T00:00:00.000Z
                    process: 79
                  - name: Ut in oc
                    deadline: 1953-02-02T00:00:00.000Z
                    planedDate: 1952-11-30T00:00:00.000Z
                    description: tempor ex in
                    importance: 3
                    entererAt: 1978-04-24T00:00:00.000Z
                    process: 74
                  - name: non ni
                    deadline: 1944-08-03T00:00:00.000Z
                    planedDate: 1962-04-18T00:00:00.000Z
                    description: quis voluptate sed anim
                    importance: 5
                    entererAt: 1954-12-24T00:00:00.000Z
                    process: 8
              schema:
                items:
                  $ref: '#/components/schemas/Task'
                type: array
          description: All Tasks of the user in an list
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Get All Tasks
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
    post:
      description: Adds a task to the users storage
      operationId: addTask
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/InitialTask'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: |-
            task added.
            The whole created Task in response
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Add Task
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
      x-contentType: application/json
  /user/{username}/task/{task}:
    delete:
      description: Deletes the task
      operationId: deleteTask
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The name of the task to edit delte or get detailed information
        explode: false
        in: path
        name: task
        required: true
        schema:
          maxLength: 50
          minLength: 1
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    name: esse con
                    deadline: 1984-07-28T00:00:00.000Z
                    planedDate: 1991-01-19T00:00:00.000Z
                    description: deserunt
                    importance: 8
                    entererAt: 2009-06-28T00:00:00.000Z
                    process: 8
              schema:
                $ref: '#/components/schemas/Task'
          description: data delivered
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: task not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The Rask you need is not avainable, mabye you create it first?
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
    get:
      description: get detailed information about the task
      operationId: getTask
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The name of the task to edit delte or get detailed information
        explode: false
        in: path
        name: task
        required: true
        schema:
          maxLength: 50
          minLength: 1
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    name: esse con
                    deadline: 1984-07-28T00:00:00.000Z
                    planedDate: 1991-01-19T00:00:00.000Z
                    description: deserunt
                    importance: 8
                    entererAt: 2009-06-28T00:00:00.000Z
                    process: 8
              schema:
                $ref: '#/components/schemas/Task'
          description: data delivered
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: task not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The Rask you need is not avainable, mabye you create it first?
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
    patch:
      description: Edit the Task
      operationId: editTask
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The name of the task to edit delte or get detailed information
        explode: false
        in: path
        name: task
        required: true
        schema:
          maxLength: 50
          minLength: 1
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/EditTask'
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    name: esse con
                    deadline: 1984-07-28T00:00:00.000Z
                    planedDate: 1991-01-19T00:00:00.000Z
                    description: deserunt
                    importance: 8
                    entererAt: 2009-06-28T00:00:00.000Z
                    process: 8
              schema:
                $ref: '#/components/schemas/Task'
          description: |-
            EDITED.
            The edited task
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: task not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The Rask you need is not avainable, mabye you create it first?
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - tag: Tasking
      x-contentType: application/json
  /user/{username}/group/{group_UUID}/tasks:
    delete:
      description: Deletes all Tasks
      operationId: removeAllGroupTasks
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The UUID of the group to operate with
        explode: false
        in: path
        name: group_UUID
        required: true
        schema:
          maxLength: 30
          minLength: 30
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: all tasks deleted
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: all tasks deleted
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        403:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: no permission
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The user doesn't has the permission to perform this acction
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Remove All Group Tasks
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Tasking
      - hasMore: "true"
        tag: Grouping
      - tag: MemberingGroups
    get:
      description: Fetch the users tasks
      operationId: getAllGroupTasks
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The UUID of the group to operate with
        explode: false
        in: path
        name: group_UUID
        required: true
        schema:
          maxLength: 30
          minLength: 30
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                  - name: do
                    deadline: 1944-12-07T00:00:00.000Z
                    planedDate: 2010-08-11T00:00:00.000Z
                    description: minim sunt in do culpa
                    importance: 0
                    entererAt: 1988-10-06T00:00:00.000Z
                    process: 79
                  - name: Ut in oc
                    deadline: 1953-02-02T00:00:00.000Z
                    planedDate: 1952-11-30T00:00:00.000Z
                    description: tempor ex in
                    importance: 3
                    entererAt: 1978-04-24T00:00:00.000Z
                    process: 74
                  - name: non ni
                    deadline: 1944-08-03T00:00:00.000Z
                    planedDate: 1962-04-18T00:00:00.000Z
                    description: quis voluptate sed anim
                    importance: 5
                    entererAt: 1954-12-24T00:00:00.000Z
                    process: 8
              schema:
                items:
                  $ref: '#/components/schemas/Task'
                type: array
          description: All Tasks of the user in an list
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Get All Group Tasks
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Tasking
      - hasMore: "true"
        tag: Grouping
      - tag: MemberingGroups
    post:
      description: Adds a task to the users storage
      operationId: addGroupTask
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The UUID of the group to operate with
        explode: false
        in: path
        name: group_UUID
        required: true
        schema:
          maxLength: 30
          minLength: 30
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/InitialTask'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: |-
            task added.
            The whole created Task in response
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        403:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: no permission
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The user doesn't has the permission to perform this acction
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Add Group Task
      tags:
      - Tasking
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Tasking
      - hasMore: "true"
        tag: Grouping
      - tag: MemberingGroups
      x-contentType: application/json
...
components:
...
  requestBodies:
    InitialTask:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InitialTask'
      x-examples:
        application/json:
          name: L
          deadline: 1988-03-20T00:00:00.000Z
          planedDate: 1974-09-03T00:00:00.000Z
          description: pariatur
          importance: 4
...
    EditTask:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EditTask'
      description: Only need to enter the values to edit
      x-examples:
        application/json:
          deadline: 1966-06-25T00:00:00.000Z
          planedDate: 2009-07-03T00:00:00.000Z
          description: qui eu
          importance: 10
...
  schemas:
    Task:
      additionalProperties: false
      description: A task is a thing you need to do till a certain point.
      example:
        process: 8
        importance: 4
        name: name
        description: description
        entererAt: 2000-01-23
        deadline: 2000-01-23
        planedDate: 2000-01-23
      properties:
        name:
          description: |-
            The name of the Task.
            needs to be uniqe
          maxLength: 50
          minLength: 1
          type: string
        deadline:
          description: The last date where it is totaly needs to be done.
          format: date
          type: string
        planedDate:
          description: The date at which you planed to be done with the task
          format: date
          pattern: yyyy-mm-dd
          type: string
        description:
          description: The description of the task
          type: string
        importance:
          description: |-
            A google playstore like rating.
            it is from zero to ten. Each point describing a half star.
            10 is the best as it is 5 Stars
          example: 4
          format: int32
          maximum: 10
          minimum: 0
          multipleOf: 1
          type: integer
        entererAt:
          description: The date the taks was created
          format: date
          pattern: yyyy-mm-dd
          type: string
        process:
          description: How far the task is done in percent
          format: int32
          maximum: 100
          minimum: 0
          type: integer
      required:
      - deadline
      - description
      - entererAt
      - importance
      - name
      - planedDate
      - process
      type: object
...
    InitialTask:
      additionalProperties: false
      description: A reduced task for adding
      example:
        importance: 4
        name: name
        description: description
        deadline: 2000-01-23
        planedDate: 2000-01-23
      properties:
        name:
          description: |-
            The name of the Task.
            needs to be uniqe
          maxLength: 50
          minLength: 1
          type: string
        deadline:
          description: The last date where it is totaly needs to be done.
          format: date
          type: string
        planedDate:
          description: The date at which you planed to be done with the task
          format: date
          type: string
        description:
          description: The description of the task
          type: string
        importance:
          description: |-
            A google playstore like rating.
            it is from zero to ten. Each point describing a half star.
            10 is the best as it is 5 Stars
          example: 4
          format: int32
          maximum: 10
          minimum: 0
          multipleOf: 1
          type: integer
      required:
      - deadline
      - description
      - importance
      - name
      - planedDate
      type: object
...
    ErrorResponse:
      description: Detailed information about an Error
      properties:
        message:
          description: A short description only a few words, what went wrong
          enum:
          - api key missing
          - api key wrong
          - group not found
          - group not owned
          - no password
          - no permission
          - user not found
          - params not matching pattern
          - No email address
          - wrong password
          - User allready exists
          - task_not_found
          type: string
      required:
      - message
      type: object
    SuccessResponse:
      description: The Request was successfull handled
      example:
        message: registration worked
      properties:
        message:
          description: A short mesage what went good
          enum:
          - registration worked
          - all tasks deleted
          type: string
      required:
      - message
      type: object
...
  securitySchemes:
    User_Key:
      in: header
      name: user_key
      type: apiKey